Updated validation functions and file names
