+++ client/src/pages/Home/Home.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import * as actions from '../../actions';
-import { homeValidation } from '../../utils/validation';
+import validate from '../../utils/validation';
 import logo from '../../images/kgnu_logo.png';
 
 import { Link, SubmitBtn } from '../../components/Button';
@@ -25,18 +25,20 @@ class Home extends Component {
   handleInputChange = e => {
     const { name, value } = e.target;
 
-    homeValidation.input(e.target);
-
     this.setState({
       [name]: value
     });
+  }
+
+  handleInputBlur = e => {
+    validate.input(e.target);
   }
 
   handleFormSubmit = e => {
     e.preventDefault();
 
     const { email, password } = this.state;
-    const valid = homeValidation.submit();
+    const valid = validate.submit();
 
     if (valid) {
       this.props.loginUser({ email, password }, () => {
@@ -67,6 +69,7 @@ class Home extends Component {
                   <FormInput
                     name="email"
                     onChange={this.handleInputChange}
+                    onBlur={this.handleInputBlur}
                     type="text"
                     value={this.state.email}
                   />
@@ -78,6 +81,7 @@ class Home extends Component {
                   <FormInput
                     name="password"
                     onChange={this.handleInputChange}
+                    onBlur={this.handleInputBlur}
                     type="password"
                     value={this.state.password}
                   />
